Remove visibleWhen of menu hanging from org.dice.menu.tools (Fixes #3)
Submenu of org.dice.menu.tools will be "always" shown (but in practice,
it only will be shown in DICE RCP, firing a warning in standalone mode).

Also, a new "DICE Simulation" menu that only shows when the Papyrus
editor is open and we're not in the DICE RCP has been added. This
condition makes that both menus do not collide.