Manually register the provider to avoid shadowing non-testing providers
We provide a method to manually register the connection provider to
manually do it when we launch the execution of the test. We do it to
avoid shadowing other providers (such as the SshUiProvider) when
debugging the DICE Simulation Tool while the tests plugin is in the
development workspace.

In any case, if in the future this procedure becomes problematic, we can
simply register the testConnectionProvider at startup but with the
lowest priority, and, if developers want to make use of it, they can
simply increase the priority to the max value by themselves.